requirements.txt
websockets>=12.0
pandas>=2.0.0
numpy>=1.24.0
uvloop>=0.17.0
plotly>=5.15.0
dash>=2.14.0
python-telegram-bot>=20.0
PyYAML>=6.0
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pytz>=2023.3
